feat: add autocomplete files and keyboard module
- Added `autocomplete_en.txt` and `autocomplete_es.txt` to the native module configuration.
- Updated `package.json` to include new dependencies: `expo-notifications` and `react-native-device-info`.
- Modified `ClipboardScreen.tsx` to handle undefined search text.
- Introduced `KeyboardModule.ts` for keyboard interactions with default implementations.
- Enhanced notification constants to include `loggedInStatusChannel`.
- Updated translations for English and Spanish to include new messages.
- Refactored app build scripts to improve configuration handling.
- Added checks to prevent backup and data upload updates in development mode.
- Updated versioning in `data.json` for consistent versioning across platforms.
- Improved type definitions for notifications and translations.

+++ server/database/backups/index.ts
@@ -87,6 +87,8 @@ export const decryptFile = (filePath: string, password: string) => {
  * ```
  */
 export const handleBackupDatabase = async () => {
+  if (["1", "true"].includes(env.__DEV__)) return;
+
   await deletePreviousBackups();
   const client = await pool.connect();
   try {
